services/llm-svc/pyproject.toml
[project]
name = "llm-svc"
version = "2.0.0"
description = "LLM Analysis Service — gRPC-Microservice fuer textuelle KI-Analyse von UC-Panel-Daten"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "TI-Radar Team" },
]

dependencies = [
    # LLM Providers
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-generativeai>=0.8.0",
    # gRPC
    "grpcio>=1.68.0",
    "grpcio-tools>=1.68.0",
    "grpcio-health-checking>=1.68.0",
    "grpcio-reflection>=1.68.0",
    "protobuf>=5.29",
    # Konfiguration
    "pydantic-settings>=2.6.0",
    # Observability
    "prometheus-client>=0.21.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
